pageindex: replaced row/col grid with css-colums, restyled letterboxes
This fixes #80, but there is still room for improvement.

A major change for the page index is, that the order of pages in changed
from horizontal first to vertical first. e.g. from

A B C    A D G
D E F -> B E H
G H I    C F I

which is as one would expect it in a book.