added alpine based "-slim" Docker image
In the container based upon the -slim image the uwsgi will be run
as unprivileded www-data user. Due to using alpine the image is
smaller in size, due to neither running nginx nor supervisord
the amount of consumed memory is lower. The published port changed
from 80 to 8080.

This was discussed in #121.