Added auto save (as draft) feature to the editor. Closes #66.
When a user starts editing a page, up to every five seconds the content
will we send as draft to the server. The server stores the draft (for
anonymous users only in the session). Is the page saved in the process,
the draft is discarded. If the page is not saved, the user will be
prompted if the editing of the drat should be continued or the draft
should be discarded.