feat: added custom sidebar menu, configurable via SIDEBAR_CUSTOM_MENU
The menu can be configured both via the setting SIDEBAR_CUSTOM_MENU,
where the config is stored as json list of lists, e.g.

    [["Example",""], ["Pagetitle","Subfolder/Pagetitle"], ["External
    Link", "https://example.com""]]

In the admin preferences the Menu can be configured in a convenient way. For
adding wiki pages a dropdown select box with all pages is available. The
order can be sorted manually via drag and drop.

This was discussed in #125.