feature: Added a convenient way to add attachments and wikilinks.
As brought up in #128, a more convient way of adding links to existing
pages was missing in the editor. The right sidebar, which was the home
of un-convenient copy-n-paste markdown codes for attachments, has been
reworked: Now pages and attachments can be picked from a dropdown list
and inserted at the current cursor position in the editor.