Show the created by info in theauthorsignature plugin
Add a new line at the footer to include the Created <datetime> by <author>
of the first commit of the file. This will bring differentiation from the
last update commit author and the original author of the page.

Note that this approach will miss if the file was moved as the creator
will now be the one that moved the file and not the one that created it.