feat: OtterWikiPluginSpec got the setup() hook
This hooks enables plugins to have access to An Otter Wiki core objects:
the app, the database and the storage.

The example plugin plugin_authorsignature has been refactored to make
use of the storage object and is now able to fetch the first author
from the storage and can display it next to the last committer.